Remove Constants.java
